mmockito/mock: Create mock() class
Create core mock() class, using the same general structure as in the
feasibility study: there is a "mockery" struct which holds all the
mocked functions. The class then overrides subsref, and adds to the
mockery struct if the correct arguments are passed. It currently does
not actually return the mocked functions and there is not much error
checking.