mock: Use Maps instead of just indexing arrays
The previous solution would just index arrays with the appropriate
return value. This didn't work for strings (which were processed as
arrays of their numeric representations) and could never work for
structs or cells. Instead, use containers.Map to hold key-value pairs.
Map accepts strings as keys, so we convert numeric types to strings.

This still doesn't support struct or cell arguments, nor mixed argument
types (as subsref passes cells, we have to unpack the arguments with
cell2mat, which only works if all contents of the input cell array are
the same type). The solution is probably to write a custom hash function
to create strings from an arbitrary object ("stringify").