mock: Allow returning of cells
subsref wraps all calls to it with a cell: this proved a problem when
the result is supposed to be a cell itself, as cell2mat does not work on
nested cells; we were also unpacking the cell on every call. Instead,
check if we are using a cell of cells, using "iscell(S{1})" and if so,
use cellfun to apply cell2mat on each cell inside (thereby removing one
level of "cells").

Does NOT work correctly if the sub-cells are different sizes (as we set
'UniformOutput' to false, but if it's true it doesn't work at all).
There should be tests for this, for now just add a FIXME comment.