mock: Add a thenThrow keyword
Exceptions cannot be just returned using the normal mechanism, they need
to be thrown. Add a new keyword, thenThrow, to .when, which takes a
MException object as it's argument and adds it to mockery. If an
invocation is then matched to it, the exception is thrown rather than
returned.

Update the relevant acceptance test to comform to the new syntax. Now
all acceptance tests pass.