Split Invocation into Invocation and InvocationMatcher classes
Invocation had mixed responsibilities and was hard to use: split it into
two classes:

The new Invocation is simply a value class encapsulating the two
substructs which define a method call. Its constructor checks
that we are actually passed a method call and can throw an Exception if
not. The class has no other function.

InvocationMatcher takes an Invocation, disregards irrelevant data and
will in the future convert all arguments into matchers. It provides a
"matches" method which takes an Invocation parameter and returns true if
it can be matched (via matchers) to self. Due to the constructor, we are
sometimes forced to use the cumbersome InvocationMatcher(Invocation(..))
syntax, but their current relations make sense from a design
perspective.

Tests were also split into two files and updated accordingly.

NOTE: This change also disables the call to builtin in the mock class.
This was caught due to the new Invocation constructor, tests with
multiple calls on the same mock object were failing because mock
understood the second ".when.aFunc" as a function call rather than the
special "when" call. It seems there are no ill effects, but thought has
to be given if this is the right thing to do.