InvocationMatcher: Add suport for matchers
In our domain-specific language, we use the name "matchers" for
Constraints from matlab.unittest.constraints. They allow us to check if
a given method call is matched by a specific constraint, instead of
relying on an exact match. This feature is implemented in the
InvocationMatcher class; passed matchers are left as-is and constants
are converted to the IsEqualTo matcher.

Some unit tests are also added for the new behavior.