mmockito: Add Matcher interface and argThat matcher
The Constraints class provides the required features of matchers, but is
not named appropriately. Instead, make a custom abstract class Matcher
which contains just a "satisfedBy" method. This allows us to create
Matchers with appropriate names, and makes it easier for users to make
custom Matchers (just one function has to be written, as opposed to two
for Constraints).

Implement an argThat(constraint) matcher which provides an interface to
Constraints. In this way the user can leverage already written
constraints in a clear (albeit verbose) interface.

Update InvocationMatcher and tests to reflect the new interface.