Mock: Defer to builtin again
The call to builtin was temporarily removed in commit 01f919c to avoid
an error. Solve the problem correctly by wrapping the code that checks
mockery with a check if the substruct is long enough to be checked in
the first place. If the passed substruct is of length one (cannot be
checked at all) or is not found, we then defer to builtin. This allows
us to show the same error message whenever a method call isn't valid.