Mock: Add suport for "times" keyword
If "times(num)" follows a then* call, the call is only mocked a given
number of times. Usually, we mock the last statement infinitely but if
the keyword chain ends in "times", we again only mock it a given amount
of times (and then behave as if it weren't mocked).

Add acceptance tests showcasing this behavior.