Mock: Add support for 'strict' and 'tolerant' mocks
Support two modes of operation for mocks - with 'strict', calls to
functions will defer to the builtin subsref if not mocked and
potentially throw an error. With tolerant mocks, this error is instead
caught and an empty list [] is returned (representing "None" or "null").

Mocks are by default tolerant. The constructor now also accepts strings
'strict' or 'tolerant' to create the appropriate Mock.