InOrder: Add InOrder class to support in order verification
Create a new class used for in order verification. InOrder can take a
number of Mocks and will sort the calls made on them (based on invID);
it then allows simple in order verification to be performed. Not all
interactions have to be verified, but they need to have occured in
order.

Update acceptance tests to reflect the syntax and expand them to cover
multiple Mock in-order verification.