internal: Create new matcher ArgEqualTo
To support older versions, we cannot use any Constraints. Create a new
internal Matcher, ArgEqualTo, which will try to use a Constraint and
fall back to a simpler logic (supporting only primitive types) if
Constraints are not available. Change InvocationMatcher to create an
ArgEqualTo matcher, instead of directly using Constraints with
ArgThat(IsEqualTo()).

This change introduces a certain performance penalty, even when
implemented as a try/catch block. Removing support for older releases
should be considered at some point due to this.