Rename InvocationMatcher to InvocationPattern
The naming of InvocationMatcher is somewhat confusing: it too closely
resembles the class name "Matcher" and might be interpreted as a special
interface of Matcher, rather than an Invocation with more capabilities.
To reduce confusion, rename it to InvocationPattern, which more closely
describes its function: it holds a Pattern that Invocations can be
matched against.

Other considered names were MatchingInvocation and InvocationTemplate.

Also slightly update the docs of the InvocationPattern class, as they
were outdated.