Matcher: Rename "satisfiedBy" to "matches"
The name "satisfiedBy" is a holdover from the initial implementation via
matlab.unittest.constraints. As it doesn't really make sense, rename the
method to "matches"; follow-through on all existing matchers too.
Initially, find-replace made a mistake on the ArgThat matcher, which
does actually rely on constraints and needs to call "satisfiedBy".

Also rename ArgEqualTo internal property to "matcher", which accurately
describes it (instead of "constraint").